--- src/ctutor_backend/alembic/versions/d905aa2b4ff1_add_grading_improvements_status_enum_.py
@@ -1,57 +1,0 @@
-"""Add grading improvements: status enum, feedback, result reference
-
-Revision ID: d905aa2b4ff1
-Revises: fix_result_constraint_001
-Create Date: 2025-09-11 11:37:23.229786
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'd905aa2b4ff1'
-down_revision: Union[str, None] = 'fix_result_constraint_001'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('course_content', 'example_version_id',
-               existing_type=postgresql.UUID(),
-               comment=None,
-               existing_comment='DEPRECATED: Access via course_content_deployment.example_version_id instead. Will be removed in future migration.',
-               existing_nullable=True)
-    op.drop_index('idx_course_content_example_version_id', table_name='course_content')
-    op.add_column('course_submission_group_grading', sa.Column('result_id', postgresql.UUID(), nullable=True))
-    op.add_column('course_submission_group_grading', sa.Column('feedback', sa.String(length=4096), nullable=True))
-    op.alter_column('course_submission_group_grading', 'status',
-               existing_type=sa.VARCHAR(length=50),
-               type_=sa.Integer(),
-               nullable=False)
-    op.create_index('idx_grading_result', 'course_submission_group_grading', ['result_id'], unique=False)
-    op.create_foreign_key(None, 'course_submission_group_grading', 'result', ['result_id'], ['id'], onupdate='RESTRICT', ondelete='SET NULL')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'course_submission_group_grading', type_='foreignkey')
-    op.drop_index('idx_grading_result', table_name='course_submission_group_grading')
-    op.alter_column('course_submission_group_grading', 'status',
-               existing_type=sa.Integer(),
-               type_=sa.VARCHAR(length=50),
-               nullable=True)
-    op.drop_column('course_submission_group_grading', 'feedback')
-    op.drop_column('course_submission_group_grading', 'result_id')
-    op.create_index('idx_course_content_example_version_id', 'course_content', ['example_version_id'], unique=False)
-    op.alter_column('course_content', 'example_version_id',
-               existing_type=postgresql.UUID(),
-               comment='DEPRECATED: Access via course_content_deployment.example_version_id instead. Will be removed in future migration.',
-               existing_nullable=True)
-    # ### end Alembic commands ###
