+++ src/ctutor_backend/interface/results.py
@@ -7,19 +7,7 @@ from ctutor_backend.interface.base import BaseEntityGet, EntityInterface, ListQu
 from ctutor_backend.model import CourseContentType, CourseSubmissionGroupMember
 from ctutor_backend.model.course import CourseContent
 from ctutor_backend.model.result import Result
-from ctutor_backend.tasks.base import TaskStatus
-
-def map_task_status_to_int(status: TaskStatus) -> int:
-    """Map TaskStatus enum to legacy integer for database storage."""
-    mapping = {
-        TaskStatus.FINISHED: 0,    # FINISHED -> COMPLETED (0)
-        TaskStatus.FAILED: 1,      # FAILED -> FAILED (1)
-        TaskStatus.CANCELLED: 2,   # CANCELLED -> CANCELLED (2)
-        TaskStatus.QUEUED: 4,      # QUEUED -> PENDING (4)
-        TaskStatus.STARTED: 5,     # STARTED -> RUNNING (5)
-        TaskStatus.DEFERRED: 7     # DEFERRED -> PAUSED (7)
-    }
-    return mapping.get(status, 1)  # Default to FAILED (1)
+from ctutor_backend.interface.tasks import TaskStatus
 
 class ResultCreate(BaseModel):
     submit: bool
